skills/produce_w_mask/pyproject.toml
[project]
name = "Produce Product - Action Mask"
version = "0.1.0"
description = ""
authors = [{ name = "John Doe", email = "john.doe@composabl.com" }]
dependencies = [
    "composabl-core",
    "numpy"
]

[composabl]
type = "skill-teacher"
entrypoint = "produce_w_mask.teacher:BaseTeacher"
